Problems with CI's
Seems to not work properly, would greatly appreciate if someone had any comments. I think the problem lies either in the predict part (maybe when adjusting for a "averagely-gendered passenger" e.g. line 64/65 or alike for the other passenger classes) or the bootstrapping itself (I fooled around with calculating means of the CI, so some use a mean some don't)